test: more tests
